# Yet more efficient parsing of UUIDs