Below are results from my notebook:

## Before
#### Corretto 8
```
[info] Benchmark                         Mode  Cnt     Score     Error  Units
[info] UUIDBenchmarks.parseCustom       thrpt    5  1321.498 ± 142.803  ops/s
[info] UUIDBenchmarks.parseInBuiltUUID  thrpt    5   168.488 ±  26.467  ops/s
```
#### Corretto 11
```
[info] Benchmark                         Mode  Cnt     Score     Error  Units
[info] UUIDBenchmarks.parseCustom       thrpt    5  1390.546 ± 417.842  ops/s
[info] UUIDBenchmarks.parseInBuiltUUID  thrpt    5   885.639 ± 259.510  ops/s
```
#### OpenJDK 17
```
[info] Benchmark                         Mode  Cnt     Score     Error  Units
[info] UUIDBenchmarks.parseCustom       thrpt    5  1175.354 ± 234.159  ops/s
[info] UUIDBenchmarks.parseInBuiltUUID  thrpt    5  1167.585 ± 140.550  ops/s
```

## After
#### Corretto 8
```
[info] Benchmark                         Mode  Cnt     Score     Error  Units
[info] UUIDBenchmarks.parseCustom       thrpt    5  1446.843 ± 194.977  ops/s
[info] UUIDBenchmarks.parseInBuiltUUID  thrpt    5   170.562 ±  22.708  ops/s
```
#### Corretto 11
```
[info] Benchmark                         Mode  Cnt     Score     Error  Units
[info] UUIDBenchmarks.parseCustom       thrpt    5  1579.872 ± 379.859  ops/s
[info] UUIDBenchmarks.parseInBuiltUUID  thrpt    5   928.248 ± 156.684  ops/s
```
#### OpenJDK 17
```
[info] Benchmark                         Mode  Cnt     Score     Error  Units
[info] UUIDBenchmarks.parseCustom       thrpt    5  1318.247 ± 164.934  ops/s
[info] UUIDBenchmarks.parseInBuiltUUID  thrpt    5  1073.430 ± 155.882  ops/s
```